refactor: 양반향 의존되있는 ordertable,order DIP 적용
